Consider person state in census side
I'm using the "pending" authorization concept in decidim for this. If
the authorization is pending in the census sie, an "pending
authorization" modal pops up. Clickin on the "check authorization
progress" button takes you to the "census authorization account page",
where you can see the submitted data, and a callout informing that the
registration is being validated.